fix: allow larger jumps in hex patterns.
The limit was 65535 because 16-bits integers were used for storing the jump ranges. Now we are using 32-bits integers.

Closes #238

+++ parser/src/ast/cst2ast.rs
@@ -851,14 +851,14 @@ impl<'src> Builder<'src> {
         let mut end = None;
 
         if let Event::Token { kind: INTEGER_LIT, .. } = self.peek() {
-            let (value, _lit, _span) = self.integer_lit::<u16>()?;
+            let (value, _lit, _span) = self.integer_lit::<u32>()?;
             start = Some(value);
         };
 
         if let Event::Token { kind: HYPHEN, .. } = self.peek() {
             self.expect(HYPHEN)?;
             if let Event::Token { kind: INTEGER_LIT, .. } = self.peek() {
-                let (value, _lit, _span) = self.integer_lit::<u16>()?;
+                let (value, _lit, _span) = self.integer_lit::<u32>()?;
                 end = Some(value);
             };
         } else {
